Add rest controller

+++ src/main/java/pl/sda/springtraining/spring/user/UserRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Long> {
-    //fixme:
-    User findUserByEmail(String email);
+public interface UserRepository extends JpaRepository<User, Integer> {
+    User findUserByEmail(String email); // fixme todo
 }
